video_to_ppt_app/requirements.txt
Flask
opencv-python
python-pptx
numpy
Werkzeug
